Extract subscriptions report row helper
